LPS-141305 Move commerce logic to the proper module
