+++ modules/apps/commerce/commerce-site-initializer-extender/build.gradle
@@ -3,5 +3,17 @@ dependencies {
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
+	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:commerce:commerce-account-api")
+	compileOnly project(":apps:commerce:commerce-currency-api")
+	compileOnly project(":apps:commerce:commerce-initializer-util")
+	compileOnly project(":apps:commerce:commerce-inventory-api")
+	compileOnly project(":apps:commerce:commerce-notification-api")
+	compileOnly project(":apps:commerce:commerce-product-api")
+	compileOnly project(":apps:commerce:headless:headless-commerce:headless-commerce-admin-catalog-api")
+	compileOnly project(":apps:commerce:headless:headless-commerce:headless-commerce-admin-channel-api")
+	compileOnly project(":apps:portal-vulcan:portal-vulcan-api")
 	compileOnly project(":apps:site-initializer:site-initializer-extender:site-initializer-extender-spi")
+	compileOnly project(":core:petra:petra-function")
+	compileOnly project(":core:petra:petra-string")
 }
