+++ modules/apps/site-initializer/site-initializer-extender/site-initializer-extender/build.gradle
@@ -8,14 +8,6 @@ dependencies {
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 	compileOnly project(":apps:account:account-api")
 	compileOnly project(":apps:asset:asset-list-api")
-	compileOnly project(":apps:commerce:commerce-account-api")
-	compileOnly project(":apps:commerce:commerce-currency-api")
-	compileOnly project(":apps:commerce:commerce-initializer-util")
-	compileOnly project(":apps:commerce:commerce-inventory-api")
-	compileOnly project(":apps:commerce:commerce-notification-api")
-	compileOnly project(":apps:commerce:commerce-product-api")
-	compileOnly project(":apps:commerce:headless:headless-commerce:headless-commerce-admin-catalog-api")
-	compileOnly project(":apps:commerce:headless:headless-commerce:headless-commerce-admin-channel-api")
 	compileOnly project(":apps:document-library:document-library-api")
 	compileOnly project(":apps:dynamic-data-mapping:dynamic-data-mapping-api")
 	compileOnly project(":apps:fragment:fragment-api")
